modify: PostList 컴포넌트 세분화
- props로 들어가는 posts의 구조가 달라서, 세분화했습니다.
